Use floats instead of scaled ints for PR score (#872)
* Update PR Benchmark

* Use float instead of scaled int to keep track of page ranks

* Combine PR scores in a single pass, same as the HugePR does

* Fix LongIterators benchmark

* Ignore daming factor in HugeEigenvector PR

+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/ComputeStep.java
@@ -19,7 +19,7 @@
 package org.neo4j.graphalgo.impl.pagerank;
 
 public interface ComputeStep extends Runnable {
-    int[][] nextScores();
+    float[][] nextScores();
 
     double[] pageRank();
 
@@ -27,7 +27,7 @@ public interface ComputeStep extends Runnable {
 
     void setStarts(int[] startArray, int[] lengthArray);
 
-    void prepareNextIteration(int[][] score);
+    void prepareNextIteration(float[][] score);
 
     void prepareNormalizeDeltas(double l2Norm, int iteration);
